sound: fix 64-bit pointer truncation in the c140 and iremga20 mixers
Both mixers round-tripped real pointers through 'unsigned long', which is
32-bit under LLP64 (Win64), truncating the pointer before it was used as an
address:

  c140: pSampleData = (signed char*)((unsigned long)pRom + find_sample(...))
        -- pRom is the sample ROM base; the high half was discarded, so sample
        playback read from a bogus address. Use byte-pointer arithmetic
        directly, which is correct on every model and needs no integer cast.

  iremga20: the hand-translated asm mixer kept the ROM base and both output
        buffers in 'unsigned long' registers (edi/ebp/esi) and dereferenced
        them via *(short*)(edi + ecx) / *(char*)(esi + eax). On Win64 those
        writes/reads hit truncated addresses -> memory corruption. Make the
        three pointer-holding registers uintptr_t; the int offset registers
        (eax/ecx, including the negative-index trick) stay int and convert
        correctly. No change on ILP32/LP64.

+++ src/sound/c140.c
@@ -244,7 +244,7 @@ static void update_stereo(int ch, INT16 **buffer, int length)
 			sz=ed-st;
 
 			/* Retrieve base pointer to the sample data */
-			pSampleData=(signed char*)((unsigned long)pRom + find_sample(st,v->bank));
+			pSampleData=(signed char*)pRom + find_sample(st,v->bank);
 
 			/* Fetch back previous data pointers */
 			offset=v->ptoffset;
